Rename directory with some functional tests to `demo`
