+++ tests/functional/demo/moptconfig_random.yaml
@@ -1,0 +1,12 @@
+name: demo_random_search
+
+pool_size: 4
+max_trials: 100
+
+algorithms:
+  random
+
+database:
+  type: 'mongodb'
+  name: 'metaopt_test'
+  host: 'mongodb://user:pass@localhost'
